Set subjectAltName when generating certificate
Set 'localhost' as a subject alternative name when generating the
self-signed certificate, so that grpcurl will accept it.

Run openssl using a docker container instead of directly from the run.sh
script. The -addext flag is available only in certain versions of
openssl, so this avoids the need to have the right openssl version
installed locally.

Add a .gitignore file so the generated certificate and private key don't
show up as untracked files.